backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
motor==3.3.1
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
openai==1.3.0
pymongo==4.6.0
python-dotenv==1.0.0
langchain==0.0.335
langchain-openai==0.0.2
tiktoken==0.5.1
pypdf==3.17.0
python-docx==1.0.1
tenacity==8.2.3
